add time transformation

+++ app/src/main/java/com/amarchaud/estats/model/entity/LocationInfo.kt
@@ -12,9 +12,9 @@ data class LocationInfo(
     @ColumnInfo(name = "delta") val delta: Int = 7,
     @ColumnInfo(name = "lat") val lat: Double = 0.0,
     @ColumnInfo(name = "lon") val lon: Double = 0.0,
-    @ColumnInfo(name = "duration_day") var duration_day: Int = 0,
-    @ColumnInfo(name = "duration_week") var duration_week: Int = 0,
-    @ColumnInfo(name = "duration_month") var duration_month: Int = 0,
-    @ColumnInfo(name = "duration_year") var duration_year: Int = 0,
-    @ColumnInfo(name = "duration_all_time") var duration_all_time: Int = 0
+    @ColumnInfo(name = "duration_day") var duration_day: Long = 0, // all these value in ms
+    @ColumnInfo(name = "duration_week") var duration_week: Long = 0,
+    @ColumnInfo(name = "duration_month") var duration_month: Long = 0,
+    @ColumnInfo(name = "duration_year") var duration_year: Long = 0,
+    @ColumnInfo(name = "duration_all_time") var duration_all_time: Long = 0
 )
